+++ migrations/Version20250503210024.php
@@ -22,6 +22,7 @@ final class Version20250503210024 extends AbstractMigration
         $this->addSql('CREATE TABLE categories (
             id INT AUTO_INCREMENT NOT NULL, 
             name VARCHAR(255) NOT NULL, 
+            is_default TINYINT(1) NOT NULL, 
             PRIMARY KEY(id)) DEFAULT CHARACTER SET utf8mb4 COLLATE `utf8mb4_unicode_ci` ENGINE = InnoDB'
         );